requirements.txt
feedparser
jinja2
requests
beautifulsoup4
psycopg2-binary
